Update README.md
a fix for what I assume was a typo
explanation - `{first, last}` desugars to `{first: first, last: last}`, but first and last aren't defined anywhere (and because of js quirks, will think it means `window.first` and `window.last` as the values). this fix removes the shorthand to be more explicit and avoids the bug.